Moved all static header files into include directory
