+++ include/cyberprobe/protocol/tls_exception.h
@@ -11,13 +11,15 @@
 #include <cyberprobe/exception.h>
 
 namespace cyberprobe {
+namespace protocol {
 
-// A tls processing exception
+    // A tls processing exception
     class tls_exception : public exception {
     public:
         tls_exception(const std::string& m) : exception(m) {}
     };
 
+}
 }
 
 #endif
