+++ include/cyberprobe/protocol/tls_utils.h
@@ -1,0 +1,14 @@
+
+#ifndef TLS_UTILS_H
+#define TLS_UTILS_H
+
+#include <string>
+
+namespace cyberprobe {
+    namespace protocol {
+        namespace tls_utils {
+            std::string convertTLSVersion(uint8_t maj, uint8_t min);
+        }
+    }
+}
+#endif
