+++ src/address.C
@@ -1,8 +1,7 @@
 
 #include <cyberprobe/network/socket.h>
 #include <cyberprobe/protocol/address.h>
-
-#include "hardware_addr_utils.h"
+#include <cyberprobe/util/hardware_addr_utils.h>
 
 #include <iomanip>
 #include <arpa/inet.h>
@@ -145,7 +144,7 @@ void address::get(std::string& cls, std::string& address) const
 
     if (proto == WLAN) {
         cls = "802.11";
-        address = hw_addr_utils::to_string(&addr[0]);
+        address = util::hw_addr_utils::to_string(&addr[0]);
         return;
     }
