--- src/hardware_addr_utils.h
@@ -1,16 +1,0 @@
-
-#ifndef HARDWARE_ADDR_UTILS_H
-#define HARDWARE_ADDR_UTILS_H
-
-#include <string>
-
-namespace cyberprobe {
-    namespace hw_addr_utils {
-
-        // no bounds checking, must be done in caller
-        std::string to_string(const uint8_t* addr);
-
-    }
-}
-
-#endif
